+++ src/store/index.js
@@ -25,13 +25,7 @@ const message = (state = initialState, action) => {
         case FETCH:
             return (state = action.value);
         case ADD_MESSAGE:
-            return state.concat({
-                userId: action.userID,
-                userName: action.userName,
-                profileImage: action.profileImage,
-                content: action.content,
-                date: new Date(),
-            });
+            return (state = action.value);
         case REMOVE_MESSAGE:
             return state.filter((item) => item.date !== action.date);
         default:
